+++ R/response.R
@@ -391,6 +391,22 @@ Response <- R6::R6Class(
       assert_that(not_missing(name))
       assert_that(not_missing(value))
 
+      if(length(.globals$cookiePreprocessors) > 0) {
+        for(i in 1:length(.globals$cookiePreprocessors)) {
+          value <- .globals$cookiePreprocessors[[i]](
+            name,
+            value,
+            expires,
+            max_age,
+            domain,
+            path,
+            secure,
+            http_only,
+            same_site
+          )
+        }
+      }
+
       name <- as_label(name)
       cookie <- sprintf("%s=%s", name, value)