+++ R/zzz.R
@@ -5,4 +5,5 @@
   .globals$errorLog <- new_log(error())
   .globals$successLog <- new_log(success())
   .globals$cookieParser <- default_cookie_parser
+  .globals$cookiePreprocessors <- list()
 }